requirements.txt
stanza>=1.9.0
PyYAML>=6.0
